feat: Reorganize server wizard
