+++ lib/l10n/app_pt.arb
@@ -53,7 +53,17 @@
       }
     }
   },
-  "serverNotAddedErrorDescription": "Por favor verifique os dados inseridos e certifique-se que o servidor está online.\n\nSe você está conectando remotamente, certifique-se que as portas 7001 e 7002 estão abertas para o servidor Bluecherry!",
+  "serverNotAddedErrorDescription": "Please check the entered details and ensure the server is online.\n\nIf you are connecting remote, make sure the {port} and {rtspPort} ports are open to the Bluecherry server!",
+  "@serverNotAddedErrorDescription": {
+    "placeholders": {
+      "port": {
+        "type": "String"
+      },
+      "rtspPort": {
+        "type": "String"
+      }
+    }
+  },
   "serverAlreadyAdded": "O {serverName} servidor já foi adicionado.",
   "@serverAlreadyAdded": {
     "placeholders": {
