+++ lib/widgets/settings/mobile/settings.dart
@@ -24,13 +24,13 @@ import 'package:bluecherry_client/providers/home_provider.dart';
 import 'package:bluecherry_client/providers/server_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/providers/update_provider.dart';
+import 'package:bluecherry_client/screens/servers/edit_server.dart';
+import 'package:bluecherry_client/screens/servers/edit_server_settings.dart';
 import 'package:bluecherry_client/utils/constants.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
 import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/widgets/device_grid/device_grid.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
-import 'package:bluecherry_client/widgets/servers/edit_server.dart';
-import 'package:bluecherry_client/widgets/servers/edit_server_settings.dart';
 import 'package:bluecherry_client/widgets/settings/shared/date_language.dart';
 import 'package:bluecherry_client/widgets/settings/shared/tiles.dart';
 import 'package:bluecherry_client/widgets/settings/shared/update.dart';
